meta-nexell: meta-nexell-distro: Modify android-tools recipe
This patch fixes adb stop script.
This patch adds s5p4418_daudio_ref target adb start script.
This patch changes android-tools recipe file.

+++ meta-nexell-distro/recipes-devtools/android-tools-nexell/files/stop_adbd.sh
@@ -1,10 +1,21 @@
 #!/bin/sh
-
+GADGET_DIR=/sys/kernel/config/usb_gadget/g1
 PID_ADBD=`pidof adbd`
-kill $PID_ADBD
 
-rm /sys/kernel/config/usb_gadget/g1/configs/c.1/ffs.adb
-rmdir /sys/kernel/config/usb_gadget/g1/configs/c.1/strings/0x409
-rmdir /sys/kernel/config/usb_gadget/g1/configs/c.1
-rmdir /sys/kernel/config/usb_gadget/g1/functions/ffs.adb
-rmdir /sys/kernel/config/usb_gadget/g1/strings/0x409
+if [ -f "/run/adbd.pid" ]; then
+    kill $PID_ADBD
+    rm /run/adbd.pid
+fi
+
+if [ -d "${GADGET_DIR}/functions/ffs.adb" ]; then
+    rm ${GADGET_DIR}/configs/c.1/ffs.adb
+    rmdir ${GADGET_DIR}/configs/c.1/strings/0x409
+    rmdir ${GADGET_DIR}/configs/c.1
+    rmdir ${GADGET_DIR}/functions/ffs.adb
+    rmdir ${GADGET_DIR}/strings/0x409
+fi
+
+if [ -d "/dev/usb-ffs/adb" ]; then
+    umount /dev/usb-ffs/adb
+    rmdir /dev/usb-ffs/adb
+fi
